data/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.29.0
pandas==2.2.1
psycopg2==2.9.9
python-dateutil==2.9.0.post0
SQLAlchemy==2.0.28